Refactor: useRegister hook 분리
